refactor: extract AdminNamesTab subcomponents
