+++ src/features/dashboard/components/admin/components/AdminNamesTypes.ts
@@ -1,0 +1,8 @@
+import type { NameItem } from "@/shared/types";
+
+export type NameWithStats = NameItem & {
+	votes?: number;
+	popularityScore?: number;
+};
+
+export type BulkAction = "hide" | "unhide" | "lock" | "unlock";
